Create AwwImage class and move object formation to Repo
Create new class to hold all relevant info from API call. This cuts down
on need for parcelable classes and consolidates formation. It also
allows for relevant info to be easily saved and reused later.

Also implement actions "share link" and "go to link" in image activity.

Also fix picasso to fit images to screen and center inside and add
error drawable and a placeholder for the placeholder.